// api/model/plandetails.model.ts
export class AveragePrice {

  kwh500: string;
  kwh1000: string;
  kwh2000: string;
}

export class Promo {

    publishDate: string;
    priority: string;
    offerName: string;
    offerTagline: string;
    signUpBonus: string;
    marketingText: string;
    disclaimer: string;
    learnMore: string;
    averagePrice: AveragePrice;
    energyCharge: string;
    baseCharge: string;
    tdspCharge: string;
    planType: string;
    contractTerm: string;
    cancellationFee: string;
}